Added controllable player objects that can fire impulses
P1 controls with arrows and fires with right control.
P2 controls with wasd and fires with left control.
Impulses have a cooldown (editable with the field)

Note: Player 1 must be tagged Player1Owned and player 2 Player2Owned respectively.

If Player collision doesn't work check the layers your objects and walls are on.